Convert int array to BYTE array, saves space

+++ bdos/time.c
@@ -63,7 +63,7 @@ extern void (*etv_timer)(int);
 
 static void tikfrk(int n);
 
-static const int nday[] = {0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
+static const BYTE nday[] = {0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
 
 /* static long uptime; */
